+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from "@angular/core";
-import { Store } from "@ngrx/store";
 import { Observable } from "rxjs";
 
 import { Product } from "../../models/product";
@@ -8,6 +7,8 @@ import { CartService } from "../../../cart/services/cart.service";
 import { selectProducts } from "../../store/products.selectors";
 import * as ProductActions from '../../store/products.actions';
 import * as CartActions from "../../../cart/store/cart.actions";
+import { ProductsFacade } from "../../services/products.facade";
+import { CartFacade } from "src/app/cart/services/cart.facade";
 
 @Component({
     selector: 'app-product-list',
@@ -16,26 +17,34 @@ import * as CartActions from "../../../cart/store/cart.actions";
   })
 export class ProductListComponent implements OnInit
 {
-    readonly products$: Observable<Product[]> = this.store.select(selectProducts);
+    //readonly products$: Observable<Product[]> = this.store.select(selectProducts);
+    readonly products$: Observable<Product[]> = this.productsFacade.products$;
 
     // constructor(private productsService: ProductsService, private cartService: CartService) {
     // }
 
-    constructor(private store: Store,  private cartService: CartService) {
+    constructor(private productsFacade: ProductsFacade,  private cartFacade: CartFacade) {
     }
 
     ngOnInit(): void {
-        this.store.dispatch(ProductActions.opened());
+        this.productsFacade.loadProducts();
+        //this.store.dispatch(ProductActions.opened());
         //this.products$ = this.productsService.getProducts();
     }
 
     addProduct(item: Product): void {
-        this.store.dispatch(CartActions.addItem({
+        this.cartFacade.addItem({
             id: item.id,
             name: item.name,
             price: item.price,
             amount: 1
-        }));
+        });
+        // this.store.dispatch(CartActions.addItem({
+        //     id: item.id,
+        //     name: item.name,
+        //     price: item.price,
+        //     amount: 1
+        // }));
 
         // this.cartService.addItem({
         //     id: item.id,